requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
APScheduler==3.10.4
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.1
driver==0.1.0
Gecko==1.0.19
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==23.2
pluggy==1.3.0
pyserial==3.5
PySocks==1.7.1
pytest==7.4.2
pytest-cov==4.1.0
pytz==2023.4
requests==2.31.0
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.11.1
tzlocal==5.2
urllib3==2.0.7
wsproto==1.2.0
